+++ src/dashboardWebView/state/atom/TabAtom.ts
@@ -1,7 +1,7 @@
 import { atom } from 'recoil';
 import { Tab } from '../../constants/Tab';
 
-export const TabAtom = atom<Tab>({
+export const TabAtom = atom<Tab | string>({
   key: 'TabAtom',
   default: Tab.All
 });