+++ src/helpers/ArticleHelper.ts
@@ -204,7 +204,7 @@ export class ArticleHelper {
       let newFileName = `${sanitizedName}.md`;
 
       if (prefix && typeof prefix === "string") {
-        newFileName = `${format(new Date(), DateHelper.formatUpdate(prefix))}-${newFileName}`;
+        newFileName = `${format(new Date(), DateHelper.formatUpdate(prefix) as string)}-${newFileName}`;
       }
       
       newFilePath = join(folderPath, newFileName);
@@ -251,7 +251,6 @@ export class ArticleHelper {
       const items = [{ 
         title: "Check file", 
         action: async () => {
-          console.log(fileName);
           await EditorHelper.showFile(fileName)
         } 
       }];